ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai-agents>=0.0.7
openai>=1.75.0
psycopg2-binary==2.9.9
httpx==0.27.0
beautifulsoup4==4.12.3
pydantic>=2.0
python-dotenv==1.0.1
